LiftingLine + GEBT + RigidBody

Theory

This model is defined by coupling the lifting line aerodynamics model with the geometrically exact beam theory model and the rigid body dynamics model.

Constructors

AerostructuralDynamics.couple_models — Method

couple_models(aero::LiftingLine, stru::GEBT, dyn::RigidBody)

Create an aerostructural model using a lifting line aerodynamic model coupled with a geometrically exact beam theory and rigid body model. This model introduces additional parameters corresponding to the freestream velocity components $V_x, V_y, V_z$, air density $\rho$, gravitational constant $g$, and external forces $F_{x,i}, F_{y,i}, F_{z,i}, M_{x,i}, M_{y,i}, M_{z,i}$ or displacements $u_{x,i}, u_{y,i}, u_{z,i}, \theta_{x,i}, \theta_{y,i}, \theta_{z,i}$ applied to each node.

NOTE: When using this model, the local frame for each beam element should be oriented with the x-axis along the beam's axis, the y-axis forward, and the z-axis normal to the surface

Example Initialization

using AerostructuralDynamics, GXBeam, LinearAlgebra

# discretization
N = 8 # number of elements

# geometric properties
span = 6.096 # m (wing half span)
chord = 1.8288 # m (chord)

# structural section properties
xea = 0.33*chord # m (elastic axis, from leading edge)
EIcc = 9.77e6 # N*m^2 (flat bending rigidity)
GJ = 0.99e6 # N*m^2 (torsional rigidity)
μ = 35.71 # kg/m (mass per unit length)
xcm = 0.43*chord # m (center of mass, from leading edge)
i11 = 8.64 # kg*m (moment of inertia about elastic axis)
i22 = 0.1*i11 # moment of inertia about beam y-axis
i33 = 0.9*i11 # moment of inertia about beam z-axis

# define geometry
xpt = range(0, 0, length=N+1) # point x-coordinates (in body frame)
ypt = range(0, span, length=N+1) # point y-coordinates (in body frame)
zpt = range(0, 0, length=N+1) # point z-coordinates (in body frame)
points = [[xpt[i],ypt[i],zpt[i]] for i = 1:N+1]
start = 1:N # starting point of each beam element
stop = 2:N+1 # ending point of each beam element
frames = fill([0 1 0; 1 0 0; 0 0 -1], N) # local to body frame transformation
compliance = fill(Diagonal([0, 0, 0, 1/GJ, 1/EIcc, 0]), N) # compliance matrix
xm2 = xea - xcm
mass = fill([
    μ 0 0 0 0 -μ*xm2;
    0 μ 0 0 0 0;
    0 0 μ μ*xm2 0 0;
    0 0 μ*xm2 i11 0 0;
    0 0 0 0 i22 0;
    -μ*xm2 0 0 0 0 i33], N) # mass matrix
assembly = GXBeam.Assembly(points, start, stop; frames, compliance, mass)

# boundary condition initialization
prescribed = Dict(
    # fixed left edge
    1 => GXBeam.PrescribedConditions(ux=0, uy=0, uz=0, theta_x=0, theta_y=0,
        theta_z=0),
)

aero = LiftingLine{N}(Wagner())

stru = GEBT(assembly, prescribed)

dyn = RigidBody()

model = couple_models(aero, stru, dyn)

nothing #hide