Developer Guide

This package has been designed to extensible so that custom models may be implemented and used as necessary. This guide details the process for defining and coupling new models.

Defining a New Submodel

New submodels may be defined using the following constructor:

Aeroelasticity.Submodel — Method

Submodel{iip}(f, nx, ny, np; kwargs...)

Define a custom submodel for use with Aeroelasticity.jl. Note that iip should be set to true if f is defined in-place and false otherwise.

Arguments

f: Residual function (resid = f(dx, x, y, p, t) or f(resid, dx, x, y, p, t))

nx: Number of states.

ny: Number of inputs.

np: Number of parameters.

Keyword Arguments

ratejac = Nonlinear(): Rate jacobian description (of type AbstractJacobian)

statejac = Nonlinear(): State jacobian description (of type AbstractJacobian)

inputjac = Nonlinear(): Input jacobian description (of type AbstractJacobian)

paramjac = Nonlinear(): Parameter jacobian description (of type AbstractJacobian)

tgrad = Nonlinear(): Time gradient description (of type AbstractGradient)

setstate = (states; x) -> states .= x: Function for setting rates/states using a custom set of keyword arguments.

setinput = (inputs; y) -> inputs .= y: Function for setting inputs using a custom set of keyword arguments.

setparam = (params; p) -> params .= p: Function for setting parameters using a custom set of keyword arguments.

sepstate = (x) -> (x = x,): Function for separating rates/states into named variables.

sepinput = (y) -> (y = y,): Function for separating inputs into named variables.

sepparam = (p) -> (p = p,): Function for separating parameters into named variables.

The first argument, f, is a residual function which may be used to define the submodel's state rates $\dot{x}$ as a function of the model's states $x$, inputs $y$, and parameters $p$, as well as the current time $t$. Note that for the purposes of this package we consider all time varying parameters to be inputs, rather than parameters. The function signature for this function is either f!(resid, dx, x, y, p, t) or resid = f(dx, x, y, p, t) depending on whether the submodel is specified as in-place (iip=true) or out-of-place (iip=false). The arguments dx, x, y, and p are assumed to be vectors.

The next three arguments (nx, ny, and np) define the number of state variables, inputs, and parameters respectively.

The remaining arguments are optional, but may be used to specify additional information about a model which can make the model easier to work with.

The gradients/jacobians associated with a model may be specified using the ratejac, statejac, inputjac, paramjac, and tgrad keyword arguments. Each of these arguments, if provided, must be wrapped as an object of type AbstractJacobian.

The arguments setstate, setinput, and setparam define convenience functions for setting the values of the state, input, and parameter vectors using keyword arguments.

The arguments sepstate, sepinput, and sepparam define convenience functions for separating the values of the state, input, and parameter vectors into a named tuple.

Defining a New Model Coupling

New couplings may be defined using the following constructor:

Aeroelasticity.Coupling — Method

Coupling{iip}(g, nx, ny, np, npc; kwargs...)

Define a custom coupling model for use with Aeroelasticity.jl. Note that iip should be set to true if g is defined in-place and false otherwise.

Arguments

g: Coupling function (y = g(dx, x, p, t) or g(y, dx, x, p, t))

nx: Number of states.

ny: Number of inputs.

np: Number of parameters (including parameters introduced by the coupling).

npc: Number of additional parameters introduced by the coupling.

Keyword Arguments

ratejac = Nonlinear(): Rate jacobian description (of type AbstractJacobian)

statejac = Nonlinear(): State jacobian description (of type AbstractJacobian)

paramjac = Nonlinear(): Parameter jacobian description (of type AbstractJacobian)

tgrad = Nonlinear(): Time gradient description (of type AbstractGradient)

setparam = (params; p) -> params .= p: Function for setting parameters introduced by the coupling using a custom set of keyword arguments.

sepparam = (p) -> (p = p,): Function for separating parameters introduced by the coupling into named variables.

The first argument, g, is a coupling function which may be used to define the coupled model's inputs $y$ as a function of the coupled model's state rates $dx$, states $x$ and parameters $p$, as well as the current time $t$. The function signature for this function is either g!(y, dx, x, y, p, t) or y = g(dx, x, p, t) depending on whether the model is specified as in-place (iip=true) or out-of-place (iip=false). The arguments dx, x, and y, are, and p are assumed to be vectors.

The next three arguments (nx, ny, and np) define the number of state variables, inputs, and parameters of the coupled model. The argument npc defines the number of additional parameters introduced by the coupling.

The remaining arguments are optional, but may be used to specify additional information about the coupling which can make the coupled model easier to work with.

The gradients/jacobians associated with a coupling may be specified using the ratejac, statejac, paramjac, and tgrad keyword arguments. Each of these arguments, if provided, must be wrapped as an object of type AbstractJacobian.

The argument setparam may be provided to define a convenience function for setting the values of the parameter vector associated with the coupling function using keyword arguments.

The argument sepparam may be provided to define a convenience function for separating the values of the parameter vector associated with the coupling function into a named tuple.

Defining Jacobians

The following types are available for defining gradients and jacobians:

Aeroelasticity.AbstractJacobian — Type

AbstractJacobian

Supertype for gradient/jacobian wrappers.

Aeroelasticity.Empty — Type

Empty <: AbstractJacobian

A gradient/jacobian wrapper which indicates that a given gradient/jacobian is empty.

Aeroelasticity.Zeros — Type

Zeros <: AbstractJacobian

A gradient/jacobian wrapper which indicates that a given gradient/jacobian is zero.

Aeroelasticity.Identity — Type

Identity <: AbstractJacobian

A gradient/jacobian wrapper which indicates that a given gradient/jacobian is the identity matrix.

Aeroelasticity.Invariant — Type

Invariant <: AbstractJacobian

A gradient/jacobian wrapper which indicates that the value of given gradient/jacobian is independent of the state rates, states, inputs, parameters, and time.

The value of the invariant jacobian/gradient may be provided upon construction, otherwise it will be calculated using automatic differentiation.

Aeroelasticity.Constant — Type

Constant <: AbstractJacobian

A gradient/jacobian wrapper which indicates that the value of given gradient/jacobian is independent of the state rates, states, inputs, and time.

The jacobian/gradient will be calculated using automatic differentiation. Alternatively, a function may be provided which defines the value of the gradient/jacobian as a function of the parameters.

Aeroelasticity.Linear — Type

Linear <: AbstractJacobian

A gradient/jacobian wrapper which indicates that the value of given gradient/jacobian is linear with respect to the differentiated variable.

The jacobian/gradient will be calculated using automatic differentiation. Alternatively, a function may be provided which defines the value of the gradient/jacobian as a function of the state rates, states, inputs, parameters, and time.

Aeroelasticity.Nonlinear — Type

Nonlinear <: AbstractJacobian

A gradient/jacobian wrapper which indicates that the value of given gradient/jacobian is nonlinear with respect to the differentiated variable.

The jacobian/gradient will be calculated using automatic differentiation. Alternatively, a function may be provided which defines the value of the gradient/jacobian as a function of the state rates, states, inputs, parameters, and time.

In general, the most specific jacobian specification should be used whenever possible, though this is most important when defining ratejac.

Explicit Models

Residual functions for explicit models, (i.e. models whose governing equations can be written as $\dot{x} = f(x, y, p, t)$) should be written in the form $0 = \dot{x} - f(x, y, p, t)$ when used with this package. Setting ratejac = Identity() will then allow this package to recover the explicit relationship from the specified residual function.

Defining a New Coupled Model

Custom coupled models may be created by using the assemble_model function. This function uses the following order when concatenating the states, inputs and parameters of a coupled model:

Aerodynamic Model

Structural Model

Dynamics Model

Control Surface Model

Controller Model

Coupling Model