Interface

User Interface

AerostructuralDynamics.AbstractModel — Type

AbstractModel

Supertype for all models.

AerostructuralDynamics.number_of_states — Function

number_of_states(models)

Return the total number of states corresponding to the model or models.

AerostructuralDynamics.number_of_inputs — Function

number_of_inputs(models)

Return the total number of inputs corresponding to the model or models.

AerostructuralDynamics.number_of_parameters — Function

number_of_parameters(models)

Return the total number of parameters corresponding to the model or models.

AerostructuralDynamics.state_indices — Function

state_indices(models)

Return the indices corresponding to the state variables for each model in models

AerostructuralDynamics.input_indices — Function

input_indices(models)

Return the indices corresponding to the input variables for each model in models

AerostructuralDynamics.parameter_indices — Function

parameter_indices(models)

Return the indices corresponding to the parameters for each model in models

AerostructuralDynamics.get_rates — Function

get_rates(models, u, y, p, t)

Calculate the (mass matrix multiplied) state rates for the specified model or models.

AerostructuralDynamics.get_rates! — Function

get_rates!(du, models, u, y, p, t)

In-place version of get_rates

AerostructuralDynamics.get_mass_matrix — Function

get_mass_matrix(models)
get_mass_matrix(models, u, y, p, t)

Calculate the mass matrix for a model or combination of models.

AerostructuralDynamics.get_mass_matrix! — Function

get_mass_matrix!(M, models)
get_mass_matrix!(M, models, u, y, p, t)

In-place version of get_mass_matrix.

AerostructuralDynamics.get_state_jacobian — Function

get_state_jacobian(models, u, y, p, t)

Calculate the jacobian with respect to the state variables for the specified models.

AerostructuralDynamics.get_state_jacobian! — Function

get_state_jacobian!(J, models, u, y, p, t)

In-place version of get_state_jacobian

AerostructuralDynamics.get_inputs — Function

get_inputs(models, u, p, t)

Calculate the inputs to the specified combination of models.

AerostructuralDynamics.get_inputs! — Function

get_inputs!(y, models::NTuple{N,AbstractModel}, u, p, t) where N

In-place version of get_inputs

AerostructuralDynamics.get_eigen — Function

get_eigen(model::TM, x, y, p, t; kwargs...)

Return the eigenvalues, left eigenvector matrix, and right eigenvector matrix corresponding to the model for state variables x, inputs y, parameters p, and time t.

For in-place models, the number of eigenvalues to compute may be specified using the nev keyword argument.

AerostructuralDynamics.get_ode — Function

get_ode(model)

Construct an ODEFunction corresponding to the specified model or models which may be solved using DifferentialEquations.

Developer Interface

AerostructuralDynamics.NoStateModel — Type

NoStateModel <: AbstractModel

Supertype for all models which contain no state variables.

AerostructuralDynamics.inplaceness — Function

inplaceness(::Type{T})

Return InPlace() if functions associated with model T are in-place or OutOfPlace() if functions associated with model T are out-of-place.

inplaceness(::Type{T1}, ::Type{T2}, ..., ::Type{TN})

Return InPlace() if the functions associated with the input function for coupled models T1, T2, ... TN are in-place or OutOfPlace() if the functions associated with the input function for coupled models T1, T2, ... TN are out-of-place.

AerostructuralDynamics.mass_matrix_type — Function

massmatrixtype(::Type{T})

Return

Empty(), if the mass matrix associated with model T is empty

Zeros(), if the mass matrix associated with model T is filled with zeros

Identity(), if the mass matrix associated with model T is the identity matrix

Constant(), if the mass matrix associated with model T is constant with respect to time

Linear(), if the mass matrix associated with model T may vary with respect to time

If no method is defined for the specified type, return Linear().

AerostructuralDynamics.state_jacobian_type — Function

statejacobiantype(::Type{T})

Return

Empty(), if the jacobian of the mass matrix multiplied state rates with respect to the state variables associated with model T is empty

Zeros(), if the jacobian of the mass matrix multiplied state rates with respect to the state variables associated with model T is filled with zeros

Identity(), if the jacobian of the mass matrix multiplied state rates with respect to the state variables associated with model T is the identity matrix

Constant(), if the jacobian of the mass matrix multiplied state rates with respect to the state variables associated with model T is constant with respect to time

Linear(), if the jacobian of the mass matrix multiplied state rates with respect to the state variables associated with model T may vary with respect to time, but is linear with respect to the states

Nonlinear(), if the jacobian of the mass matrix multiplied state rates with respect to the state variables associated with model T may vary with respect to time, and is nonlinear with respect to the states

If no method is defined for the specified type, return Nonlinear().

AerostructuralDynamics.input_jacobian_type — Function

inputjacobiantype(::Type{T})

Return

Empty(), if the jacobian of the mass matrix multiplied state rates with respect to the inputs is empty for model T

Zeros(), if the jacobian of the mass matrix multiplied state rates with respect to the inputs is filled with zeros for model T

Identity(), if the jacobian of the mass matrix multiplied state rates with respect to the inputs is the identity matrix for model T

Constant(), if the jacobian of the mass matrix multiplied state rates with respect to the inputs is constant with respect to time for model T

Linear(), if the jacobian of the mass matrix multiplied state rates with respect to the inputs may vary with respect to time for model T, and is linear with respect to the inputs

Nonlinear(), if the jacobian of the mass matrix multiplied state rates with respect to the inputs may vary with respect to time for model T, and is nonlinear with respect to the inputs

If no method is defined for the specified type, return Nonlinear().

AerostructuralDynamics.get_input_jacobian — Function

get_input_jacobian(models)
get_input_jacobian(models, u, y, p, t)

Calculate the jacobian with respect to the inputs for the specified model or models.

AerostructuralDynamics.get_input_mass_matrix — Function

get_input_mass_matrix(models)
get_input_mass_matrix(models, u, p, t)

Calculate the input function mass matrix for the specified combination of models.

AerostructuralDynamics.get_input_mass_matrix! — Function

get_input_mass_matrix!(My, models)
get_input_mass_matrix!(My, models, u, p, t)

In-place version of get_input_mass_matrix.

AerostructuralDynamics.get_input_state_jacobian — Function

get_input_state_jacobian(models::NTuple{N,AbstractModel}, u, p, t) where N

Calculate the jacobian of the input function with respect to the state variables for the specified combination of models.

AerostructuralDynamics.get_input_state_jacobian! — Function

get_input_state_jacobian!(J, models, u, p, t)

In-place version of get_input_state_jacobian