Library

AerostructuralDynamics.AbstractModel — Type

AbstractModel

Supertype for all models.

AerostructuralDynamics.Peters — Type

Peters{N,TF,SV,SA} <: AbstractModel

Peter's finite state model with N state variables, inputs $d = \begin{bmatrix} \dot{\theta} & \ddot{h} & \ddot{\theta}\end{bmatrix}^T$ and parameters $p_a = \begin{bmatrix} a & b & U & \rho \end{bmatrix}^T$

AerostructuralDynamics.Peters — Method

Peters{N,TF=Float64}()

Initialize an object of type Peters which has N aerodynamic degrees of freedom.

AerostructuralDynamics.TypicalSection — Type

TypicalSection <: AbstractModel

Typical section structural model with state variables $q = \begin{bmatrix} h & θ & \dot{h} & \dot{\theta} \end{bmatrix}^T$, structural parameters $p_s = \begin{bmatrix} a & b & k_h & k_\theta & m & x_\theta & I_P \end{bmatrix}^T$, and aerodynamic loads $r = \begin{bmatrix} L & M_\frac{1}{4} \end{bmatrix}^T$

AerostructuralDynamics.Wagner — Type

Wagner{TF} <: AbstractModel

Aerodynamic model based on Wagner's function with state variables $d = \begin{bmatrix} \lambda_1 & \lambda_2 \end{bmatrix}^T$, inputs $d = \begin{bmatrix} \theta & \dot{h} & \dot{\theta} \end{bmatrix}^T$ and parameters $p_a = \begin{bmatrix} a & b & U & \rho \end{bmatrix}^T$

AerostructuralDynamics.get_inputs! — Method

get_inputs!(y, models::NTuple{N,AbstractModel}, u, p, t) where N

In-place version of get_inputs

AerostructuralDynamics.get_inputs — Method

get_inputs(models, u, p, t)

Calculate the inputs to the specified combination of models.

AerostructuralDynamics.get_mass_matrix! — Method

get_mass_matrix!(M, models)
get_mass_matrix!(M, models, u, y, p, t)

In-place version of get_mass_matrix.

AerostructuralDynamics.get_mass_matrix — Method

get_mass_matrix(models)
get_mass_matrix(models, u, y, p, t)

Calculate the mass matrix for a model or combination of models.

AerostructuralDynamics.get_rates! — Method

get_rates!(du, models, u, y, p, t)

In-place version of get_rates

AerostructuralDynamics.get_rates — Method

get_rates(models, u, y, p, t)

Calculate the (mass matrix multiplied) state rates for the specified model or models.

AerostructuralDynamics.get_state_jacobian! — Method

get_state_jacobian!(J, models, u, y, p, t)

In-place version of get_state_jacobian

AerostructuralDynamics.get_state_jacobian — Method

get_state_jacobian(models, u, y, p, t)

Calculate the jacobian with respect to the state variables for the specified models.

AerostructuralDynamics.input_indices — Method

input_indices(models)

Return the indices corresponding to the input variables for each model in models

AerostructuralDynamics.number_of_inputs — Function

number_of_inputs(models)

Return the total number of inputs corresponding to the model or models.

AerostructuralDynamics.number_of_parameters — Function

number_of_parameters(models)

Return the total number of parameters corresponding to the model or models.

AerostructuralDynamics.number_of_states — Function

number_of_states(models)

Return the total number of states corresponding to the model or models.

AerostructuralDynamics.parameter_indices — Method

parameter_indices(models)

Return the indices corresponding to the parameters for each model in models

AerostructuralDynamics.state_indices — Method

state_indices(models)

Return the indices corresponding to the state variables for each model in models