AerostructuralDynamics.AbstractModel — Type

AbstractModel

Supertype for all models.

AerostructuralDynamics.InterfaceModel — Type

InterfaceModel <: AbstractModel

Supertype for all models which are used to extend 2D models to 3D models.

AerostructuralDynamics.LiftingLineSection — Type

LiftingLineSection <: AbstractModel

Lifting line section model with state variables $v_x, v_y, v_z, \omega_x, \omega_y, \omega_z$, inputs $F_x', F_y', F_z', M_x, M_y, M_z$, and zero parameters. Two-dimensional aerodynamic models may be extended to three-dimensional models by coupling with this model. Note that this model has no rate equations of its own since its state variables are defined as functions of the 3D structural model's state variables.

AerostructuralDynamics.LiftingLineSection — Method

LiftingLineSection()

Initialize an object of type LiftingLineSection

AerostructuralDynamics.LiftingLineSectionControl — Type

LiftingLineSectionControl <: AbstractModel

Lifting line section control input model with state variable $\delta$, zero inputs, and zero parameters. Two-dimensional control surface models may be extended to three dimensional models by coupling with this model. Note that this model has no rate equations of its own since its state variables are defined as functions of the 3D system's control inputs.

AerostructuralDynamics.NoStateModel — Type

NoStateModel <: AbstractModel

Supertype for all models which contain no state variables.

AerostructuralDynamics.ResidualModel — Type

ResidualModel <: AbstractModel

Supertype for all models whose state variables may be defined by solving a set of residual equations.

AerostructuralDynamics.UnsteadyModel — Type

UnsteadyModel <: AbstractModel

Supertype for all models which represent dynamic systems of equations.

AerostructuralDynamics.additional_parameter_indices — Method

additional_parameter_indices(models)

Return the indices of the additional parameters for the coupling function corresponding to the models in models

AerostructuralDynamics.coupling_inplaceness — Method

coupling_inplaceness(::Type{T1}, ::Type{T2}, ..., ::Type{TN})

Return InPlace() if the functions associated with the coupling function for coupled models T1, T2, ... TN are in-place or OutOfPlace() if the functions associated with the coupling function for coupled models T1, T2, ... TN are out-of-place.

AerostructuralDynamics.get_coupling_parameter_jacobian — Function

get_coupling_parameter_jacobian(model)
get_coupling_parameter_jacobian(model, p)
get_coupling_parameter_jacobian(model, dx, x, y, p, t)

Calculate the jacobian of the coupling function for model with respect to the parameters

AerostructuralDynamics.get_coupling_parameter_jacobian! — Function

get_coupling_parameter_jacobian!(J, model)
get_coupling_parameter_jacobian!(J, model, p)
get_coupling_parameter_jacobian!(J, model, dx, x, p, t)

In-place version of get_coupling_parameter_jacobian

AerostructuralDynamics.get_coupling_rate_jacobian — Function

get_coupling_rate_jacobian(model)
get_coupling_rate_jacobian(model, p)
get_coupling_rate_jacobian(model, dx, x, p, t)

Calculate the jacobian of the coupling function for model with respect to the state rates

AerostructuralDynamics.get_coupling_rate_jacobian! — Function

get_coupling_rate_jacobian!(J, model)
get_coupling_rate_jacobian!(J, model, p)
get_coupling_rate_jacobian!(J, model, dx, x, p, t)

In-place version of get_coupling_rate_jacobian.

AerostructuralDynamics.get_coupling_state_jacobian — Function

get_coupling_state_jacobian(model)
get_coupling_state_jacobian(model, p)
get_coupling_state_jacobian(model, dx, x, y, p, t)

Calculate the jacobian of the coupling function for model with respect to the state variables

AerostructuralDynamics.get_coupling_state_jacobian! — Function

get_coupling_state_jacobian!(J, model)
get_coupling_state_jacobian!(J, model, p)
get_coupling_state_jacobian!(J, model, dx, x, p, t)

In-place version of get_coupling_state_jacobian

AerostructuralDynamics.get_coupling_time_gradient — Function

get_coupling_time_gradient(model)
get_coupling_time_gradient(model, p)
get_coupling_time_gradient(model, dx, x, y, p, t)

Calculate the derivative of the coupling function for model with respect to time

AerostructuralDynamics.get_coupling_time_gradient! — Function

get_coupling_time_gradient!(dT, model)
get_coupling_time_gradient!(dT, model, p)
get_coupling_time_gradient!(dT, model, dx, x, p, t)

In-place version of get_coupling_time_gradient

AerostructuralDynamics.inplaceness — Method

inplaceness(::Type{T})

Return InPlace() if functions associated with model T are in-place or OutOfPlace() if functions associated with model T are out-of-place.

AerostructuralDynamics.input_jacobian_type — Method

input_jacobian_type(::Type{T})

Return

Empty(), if the jacobian of the residual expression associated with model T with respect to the inputs is an empty matrix

Zeros(), if the jacobian of the residual expression associated with model T with respect to the inputs is a zero matrix

Identity(), if the jacobian of the residual expression associated with model T with respect to the inputs is the identity matrix

Invariant(), if the jacobian of the residual expression associated with model T with respect to the inputs is independent of the state rates, state variables, inputs, parameters, and time.

Constant(), if the jacobian of the residual expression associated with model T with respect to the inputs is independent of the state rates, state variables, inputs, and time.

Linear(), if the jacobian of the residual expression associated with model T with respect to the inputs may vary with respect to time, and is linear with respect to the inputs.

Nonlinear(), if the jacobian of the residual expression associated with model T with respect to the inputs may vary with respect to time, and is nonlinear with respect to the inputs.

If no method is defined for the specified type, return Nonlinear().

AerostructuralDynamics.number_of_additional_parameters — Function

number_of_additional_parameters(models...)

Return the total number of additional parameters corresponding to the model coupling.

AerostructuralDynamics.parameter_jacobian_type — Method

parameter_jacobian_type(::Type{T})

Return

Empty(), if the jacobian of the residual expression associated with model T with respect to the parameters is an empty matrix

Zeros(), if the jacobian of the residual expression associated with model T with respect to the parameters is a zero matrix

Identity(), if the jacobian of the residual expression associated with model T with respect to the parameters is the identity matrix

Invariant(), if the jacobian of the residual expression associated with model T with respect to the parameters is independent of the state rates, state variables, inputs, parameters, and time.

Constant(), if the jacobian of the residual expression associated with model T with respect to the parameters is independent of the state rates, state variables, inputs, and time.

Linear(), if the jacobian of the residual expression associated with model T with respect to the parameters may vary with respect to time, and is linear with respect to the parameters.

Nonlinear(), if the jacobian of the residual expression associated with model T with respect to the parameters may vary with respect to time, and is nonlinear with respect to the parameters.

If no method is defined for the specified type, return Nonlinear().

AerostructuralDynamics.rate_jacobian_type — Method

rate_jacobian_type(::Type{T})

Return

Empty(), if the jacobian of the residual expression associated with model T with respect to the state rates is an empty matrix

Zeros(), if the jacobian of the residual expression associated with model T with respect to the state rates is a zero matrix

Identity(), if the jacobian of the residual expression associated with model T with respect to the state rates is the identity matrix

Invariant(), if the jacobian of the residual expression associated with model T with respect to the state rates is independent of the state rates, state variables, inputs, parameters, and time.

Constant(), if the jacobian of the residual expression associated with model T with respect to the state rates is independent of the state rates, state variables, inputs, and time.

Linear(), if the jacobian of the residual expression associated with model T with respect to the state rates may vary with respect to time, and is linear with respect to the state rates.

Nonlinear(), if the jacobian of the residual expression associated with model T with respect to the state rates may vary with respect to time, and is nonlinear with respect to the state rates.

If no method is defined for the specified type, return Nonlinear().

AerostructuralDynamics.separate_additional_parameters — Function

separate_additional_parameters(model, p)

Separate the additional parameter vector entries in p corresponding to the model coupling, if applicable.

AerostructuralDynamics.state_jacobian_type — Method

state_jacobian_type(::Type{T})

Return

Empty(), if the jacobian of the residual expression associated with model T with respect to the state variables is an empty matrix

Zeros(), if the jacobian of the residual expression associated with model T with respect to the state variables is a zero matrix

Identity(), if the jacobian of the residual expression associated with model T with respect to the state variables is the identity matrix

Invariant(), if the jacobian of the residual expression associated with model T with respect to the state variables is independent of the state rates, state variables, inputs, parameters, and time.

Constant(), if the jacobian of the residual expression associated with model T with respect to the state variables is independent of the state rates, state variables, inputs, and time.

Linear(), if the jacobian of the residual expression associated with model T with respect to the state variables may vary with respect to time, and is linear with respect to the states.

Nonlinear(), if the jacobian of the residual expression associated with model T with respect to the state variables may vary with respect to time, and is nonlinear with respect to the states.

If no method is defined for the specified type, return Nonlinear().

AerostructuralDynamics.time_gradient_type — Method

time_gradient_type(::Type{T})

Return

Empty(), if the derivative of the residual expression associated with model T with respect to time is an empty matrix

Zeros(), if the derivative of the residual expression associated with model T with respect to time is a zero matrix

Invariant(), if the derivative of the residual expression associated with model T with respect to time is independent of the state rates, state variables, inputs, parameters, and time.

Constant(), if the derivative of the residual expression associated with model T with respect to time is independent of the state rates, state variables, inputs, and time.

Linear(), if the derivative of the residual expression associated with model T with respect to time may vary with respect to time, and is linear with respect to time.

Nonlinear(), if the derivative of the residual expression associated with model T with respect to time may vary with respect to time, and is nonlinear with respect to time.

If no method is defined for the specified type, return Nonlinear().