Public Documentation

Documentation for AerostructuralDynamics.jl's public interface.

See the Models section of the manual for documentation covering the individual models.

Contents

Public Documentation

Contents

Index

Public Interface

Index

AerostructuralDynamics.get_coupling_inputs

AerostructuralDynamics.get_coupling_inputs!

AerostructuralDynamics.get_eigen

AerostructuralDynamics.get_ode

AerostructuralDynamics.get_state_jacobian

AerostructuralDynamics.get_state_jacobian!

AerostructuralDynamics.input_indices

AerostructuralDynamics.number_of_inputs

AerostructuralDynamics.number_of_parameters

AerostructuralDynamics.number_of_states

AerostructuralDynamics.parameter_indices

AerostructuralDynamics.state_indices

Public Interface

Getting Model Properties

The following methods may be used to determine the properties of a model.

AerostructuralDynamics.number_of_states — Function

number_of_states(model)

Return the total number of state variables corresponding to the model or models.

AerostructuralDynamics.number_of_inputs — Function

number_of_inputs(model)

Return the total number of inputs corresponding to the model or models.

AerostructuralDynamics.number_of_parameters — Function

number_of_parameters(model)

Return the total number of parameters corresponding to the model or models.

Getting Model Indices

The following methods may be used to find the state, input, and parameter indices associated with each model in a coupled model.

AerostructuralDynamics.state_indices — Function

state_indices(models)

Return the indices corresponding to the state variables for each model in models

AerostructuralDynamics.input_indices — Function

input_indices(models)

Return the indices corresponding to the input variables for each model in models

AerostructuralDynamics.parameter_indices — Function

parameter_indices(models)

Return the indices corresponding to the parameters for each model in models

Getting Model Inputs

The following methods may be used to calculate the value of the inputs for a coupled model.

AerostructuralDynamics.get_coupling_inputs — Function

get_coupling_inputs(models, dx, x, p, t)

Calculate the inputs to the specified combination of models.

AerostructuralDynamics.get_coupling_inputs! — Function

get_coupling_inputs!(y, models::NTuple{N,AbstractModel}, dx, x, p, t) where N

In-place version of get_coupling_inputs

Getting State Rates

The following methods may be used to calculate the value of the state rates for a model.

Missing docstring.

Missing docstring for get_rates. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_rates!. Check Documenter's build log for details.

Getting Mass Matrices

The following methods may be used to calculate the mass matrix for a model.

Missing docstring.

Missing docstring for get_mass_matrix. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_mass_matrix!. Check Documenter's build log for details.

Getting Jacobians

The following methods may be used to calculate the jacobian of the state rates for a model.

AerostructuralDynamics.get_state_jacobian — Function

get_state_jacobian(model)
get_state_jacobian(model, p)
get_state_jacobian(model, dx, x, y, p, t)

Calculate the jacobian of the residual function for model with respect to the state variables

AerostructuralDynamics.get_state_jacobian! — Function

get_state_jacobian!(J, model)
get_state_jacobian!(J, model, p)
get_state_jacobian!(J, model, dx, x, y, p, t)

In-place version of get_state_jacobian

Performing a Stability Analysis

The following method may be used to perform a stability analysis for a model.

AerostructuralDynamics.get_eigen — Function

get_eigen(model::TM, x, p; kwargs...)

Return the eigenvalues, left eigenvector matrix, and right eigenvector matrix corresponding to the model for state variables x and parameters p.

For in-place models, the number of eigenvalues to compute may be specified using the nev keyword argument.

Interfacing with DifferentialEquations

The following method may be used to construct a function for use with DifferentialEquations.

AerostructuralDynamics.get_ode — Function

get_ode(model)
get_ode(model, p)

Construct an ODEFunction corresponding to the specified model or models which may be solved using DifferentialEquations.