Lifting Line Model

Type Definition

Aeroelasticity.LiftingLine — Type

LiftingLine{R, Y, I}

Lifting line model constructed by concatenating the governing equations, state variables, inputs, and parameters of multiple two-dimensional aerodynamic models.

Constructors

Aeroelasticity.LiftingLine — Method

LiftingLine(fresid, nstate=number_of_states.(fresid); finput=default_coupling.(fresid, LiftingLineSection()))

Initializes a lifting line model of type LiftingLine

Arguments:

fresid: Residual function for each lifting line section

nstate: Number of state variables for each lifting line section

Keyword Arguments:

finput: Coupling function for each lifting line section.

Theory

This model is a three dimensional aerodynamic model which is constructed by concatenating the governing equations, states, inputs, and parameters of an arbitrary number of two-dimensional aerodynamic models.

Examples

Aeroelastic Analysis of the Goland Wing

Aeroelastic Analysis of a High Aspect Ratio Wing