Geometrically Exact Beam Element Assembly

Type Definition

Aeroelasticity.GXBeamAssembly — Type

GXBeamAssembly{S}

Model which describes the dynamic behavior of an assembly of beam elements, as modeled by the GXBeam package. State variables are as defined by GXBeam. Inputs are defined and passed as a GXBeamInputs struct. Parameters are defined and passed as a GXBeamParameters struct.

Missing docstring.

Missing docstring for GXBeamAssembly(). Check Documenter's build log for details.

Aeroelasticity.GXBeamInputs — Function

GXBeamInputs(; kwargs...)

Defines inputs for a geometrically exact beam theory structural model

Keyword Arguments

prescribed_conditions = Dict{Int,PrescribedConditions{Float64}}(): A dictionary with keys corresponding to the points at which prescribed conditions are applied and values of type PrescribedConditions which describe the prescribed conditions at those points.

distributed_loads = Dict{Int,DistributedLoads{Float64}}(): A dictionary with keys corresponding to the elements to which distributed loads are applied and values of type DistributedLoads which describe the distributed loads on those elements.

point_masses = Dict{Int,PointMass{Float64}}(): A dictionary with keys corresponding to the points to which point masses are attached and values of type PointMass which contain the properties of the attached point masses.

linear_velocity = zeros(3): Prescribed linear velocity of the body frame.

angular_velocity = zeros(3): Prescribed angular velocity of the body frame.

linear_acceleration = zeros(3): Prescribed linear acceleration of the body frame.

angular_acceleration = zeros(3): Prescribed angular acceleration of the body frame.

gravity = [0,0,0]: Gravity vector in the body frame.

Aeroelasticity.GXBeamParameters — Function

GXBeamParameters(assembly::GXBeam.Assembly)

Defines parameters for a geometrically exact beam theory structural model

Examples

Aeroelastic Analysis of the Goland Wing

Steady State Aeroelastic Analysis of a Highly Flexible Wing

Aeroelastic Stability Analysis of a Highly Flexible Wing

Theory

This model uses geometrically exact beam theory (as implemented by the GXBeam package to model an interconnected assembly of nonlinear beams. For more details, refer to the documentation for the GXBeam package.