Rigid Body Model

Theory

A basic six degree of freedom rigid body model based on the SixDOF package.

Type Definition

Aeroelasticity.RigidBody — Type

RigidBody(state_indices=(), rate_indices=(), prescribed_values=())

Construct a six-degree of freedom rigid-body model with state variables $xr, yr, zr, \phi, \theta, \psi, ur, vr, wr, pr, qr, rr$ and inputs $m, I_{xx}, I_{yy}, I_{zz}, I_{xz}, I_{xy}, I_{yz}, F_x, F_y, F_z, M_x, M_y, M_z$. Uses the state variables corresponding to state_indices to set the state rates corresponding to rate_indices to the values specified in prescribed_values. Otherwise, allow the state variables and their respective rates to be defined by their rate equations.

Constructors

Missing docstring.

Missing docstring for RigidBody(). Check Documenter's build log for details.

Example Initialization

model = RigidBody()