Aeroelasticity.LiftingLineGXBeamAssembly — Type

LiftingLineGXBeamAssembly()

Construct a model by coupling a lifting line aerodynamic model and a geometrically exact beam theory model (as implemented by GXBeam). Parameters for this model are defined in LiftingLineGXBeamParameters.

NOTE: When using this model, the local frame for each beam element is assumed to be oriented with the x-axis along the beam's axis, the y-axis forward (into the freestream), and the z-axis in the airfoil normal direction.

Aeroelasticity.LiftingLineSection — Type

LiftingLineSection

Lifting line section model with state variables $v_x, v_y, v_z, \omega_x, \omega_y, \omega_z$, inputs $f_x, f_y, f_z, m_x, m_y, m_z$, and parameters $\rho, c$.

Note that this model doesn't define any residuals and is defined purely as a convenience to allow aerodynamic models to be more easily used with LiftingLine. Also note that in order to preserve compatibility with the default coupling functions for the LiftingLine model, no additional parameters may be introduced when creating a coupling function involving this model.

Aeroelasticity.LiftingLineSection — Method

LiftingLineSection()

Initialize a model of type LiftingLineSection

Aeroelasticity.PetersLiftingLine — Type

PetersLiftingLine{N,TF,TV,TA}

Coupling model which allows Peters' finite state theory (see Peters) to be used more conveniently with the LiftingLine model. See LiftingLineSection.

Aeroelasticity.PetersSection — Type

PetersSection{N}

Coupling model for coupling Peters' finite state theory (see Peters) with a typical section model (see Section). This model introduces the freestream velocity $U$, air density $\rho$, and air speed of sound $c$ as additional parameters.

The parameters for the resulting coupled model (as defined by the parameter function) defaults to the parameters for each model concatenated into a single vector.

Aeroelasticity.QuasiSteadyLiftingLine — Type

QuasiSteadyLiftingLine

Coupling model which allows a quasisteady thin airfoil theory model (see Steady) to be used more conveniently with the LiftingLine model. See LiftingLineSection.

Aeroelasticity.QuasiSteadySection — Type

QuasiSteadySection

Coupling model for coupling a quasi-steady aerodynamic model based on thin airfoil theory (see QuasiSteady) and a two-degree of freedom typical section model (see [Section()]). This model introduces the freestream velocity $U$, air density $\rho$, and air speed of sound $c$ as additional parameters.

The parameters for the resulting coupled model (as defined by the parameter function) defaults to the parameters for each model concatenated into a single vector.

Aeroelasticity.SteadyLiftingLine — Type

SteadyLiftingLine

Coupling model which allows a steady thin airfoil theory model (see Steady) to be used more conveniently with the LiftingLine model. See LiftingLineSection.

Aeroelasticity.SteadySection — Type

SteadySection

Coupling model for coupling a steady aerodynamic model based on thin airfoil theory (see Steady) and a two-degree of freedom typical section model (see Section). This model introduces the freestream velocity $U_\infty$, air density $\rho_\infty$, and air speed of sound $c$ as additional parameters.

The parameters for the resulting coupled model (as defined by the parameter function) defaults to the parameters for each model concatenated into a single vector.

Aeroelasticity.WagnerLiftingLine — Type

WagnerLiftingLine{TF}

Coupling model which allows the Wagner unsteady aerodynamic model (see Wagner) to be used more conveniently with the LiftingLine model. See LiftingLineSection.

Aeroelasticity.WagnerSection — Type

WagnerSection

Coupling model for coupling an unsteady aerodynamic model based on Wagner's function (see Wagner) and a two-degree of freedom typical section model (see Section). This coupling introduces the freestream velocity $U_\infty$, air density $\rho_\infty$ and air speed of sound $c$ as additional parameters.

The parameters for the resulting coupled model (as defined by the parameter function) defaults to the parameters for each model concatenated into a single vector.

SciMLBase.DAEFunction — Method

DAEFunction(model::Aeroelasticity.CoupledModel)

Construct an DAEFunction for a coupled model.

SciMLBase.ODEFunction — Method

ODEFunction(model::Aeroelasticity.CoupledModel, p0)

Construct an ODEFunction for a coupled model. Note that by using this function you assume that the governing equations may be expressed in the form M(x, p, t)*dx = f(x, p, t)