Public Documentation

Documentation for Aeroelasticity.jl's public interface.

See the Models section of the manual for documentation covering the individual models.

Contents

Public Documentation

Contents

Index

Public Interface

Index

Aeroelasticity.get_eigen

Aeroelasticity.input_indices

Aeroelasticity.number_of_inputs

Aeroelasticity.number_of_parameters

Aeroelasticity.number_of_states

Aeroelasticity.parameter_indices

Aeroelasticity.state_indices

Public Interface

Getting Model Properties

The following methods may be used to determine the properties of a model.

Aeroelasticity.number_of_states — Function

number_of_states(model)

Return the total number of state variables corresponding to a model.

Aeroelasticity.number_of_inputs — Function

number_of_inputs(model)

Return the total number of inputs corresponding to a model.

Aeroelasticity.number_of_parameters — Function

number_of_parameters(model)

Return the total number of parameters corresponding to a model.

Getting Model Indices

The following methods may be used to find the state, input, and parameter indices associated with each model in a coupled model.

Aeroelasticity.state_indices — Function

state_indices(model::CoupledModel)

Return the state variable indices for each submodel in model

Aeroelasticity.input_indices — Function

input_indices(model::CoupledModel)

Return the input variable indices for each submodel in model

Aeroelasticity.parameter_indices — Function

parameter_indices(model::CoupledModel)

Return the parameter indices for each submodel in model

Getting Model Inputs

The following methods may be used to calculate the value of the inputs for a coupled model.

Missing docstring.

Missing docstring for get_coupling_inputs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_coupling_inputs!. Check Documenter's build log for details.

Getting State Rates

The following methods may be used to calculate the value of the state rates for a model.

Missing docstring.

Missing docstring for get_rates. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_rates!. Check Documenter's build log for details.

Getting Mass Matrices

The following methods may be used to calculate the mass matrix for a model.

Missing docstring.

Missing docstring for get_mass_matrix. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_mass_matrix!. Check Documenter's build log for details.

Getting Jacobians

The following methods may be used to calculate the jacobian of the state rates for a model.

Missing docstring.

Missing docstring for get_state_jacobian. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_state_jacobian!. Check Documenter's build log for details.

Performing a Stability Analysis

The following method may be used to perform a stability analysis for a model.

Aeroelasticity.get_eigen — Function

get_eigen(model::TM, K, M; kwargs...)

Return the eigenvalues, left eigenvector matrix, and right eigenvector matrix corresponding to the model.

For in-place models, the number of eigenvalues to compute may be specified using the nev keyword argument.

Interfacing with DifferentialEquations

The following method may be used to construct a function for use with DifferentialEquations.

Missing docstring.

Missing docstring for get_ode. Check Documenter's build log for details.