Library

AerostructuralDynamics.AbstractModel — Type

AbstractModel

Supertype for all models.

AerostructuralDynamics.GEBT — Type

GEBT <: AbstractModel

Geometrically exact beam theory model, as implemented by the GXBeam package. State variables are as defined by GXBeam. Inputs correspond to the (non-follower) distributed aerodynamic loads on each beam element with distributed loads. When coupled with an aerodynamic model, the local beam y and z-axes should be aligned with the negative chordwise and positive normal directions, respectively.

At this point in time, this model doesn't accept any parameters.

AerostructuralDynamics.GEBT — Method

GEBT(system, assembly, prescribed_conditions, distributed_loads)

Construct a geometrically exact beam theory structural model.

AerostructuralDynamics.LiftingLine — Type

LiftingLine{N,T} <: AbstractModel

Lifting line model with N cross sections, using the aerodynamic models in T. State variables, inputs, and parameters correspond to the state variables, inputs, and parameters of each of the cross sections concatenated

AerostructuralDynamics.LiftingLine — Method

LiftingLine{N}(model)

Construct a lifting line aerodynamic model using N instances of model.

AerostructuralDynamics.LiftingLine — Method

LiftingLine(models)

Construct a lifting line aerodynamic model given a tuple of aerodynamic models.

AerostructuralDynamics.LiftingLineSection — Type

LiftingLineSection <: AbstractModel

Lifting line section model with state variables $q = \begin{bmatrix} v_x & v_y & v_z & \omega_x & \omega_y & \omega_z \end{bmatrix}^T$, inputs $r = \begin{bmatrix} L' & Y' & D' & Mx, My, Mz \end{bmatrix}^T$, and no parameters. Two-dimensional aerodynamic models may be extended to three-dimensional models by coupling with this model. Note that this model has no rate equations of its own. Its state variables are defined as functions of a 3D structural model's state variables.

AerostructuralDynamics.Peters — Type

Peters{N,TF,SV,SA} <: AbstractModel

Peter's finite state model with N state variables, inputs $d = \begin{bmatrix} u & \omega & \dot{v} & \dot{\omega}\end{bmatrix}^T$ and parameters $p_a = \begin{bmatrix} a & b & a_0 & \alpha_0 \end{bmatrix}^T$

AerostructuralDynamics.Peters — Method

Peters{N,TF=Float64}()

Initialize an object of type Peters which has N aerodynamic degrees of freedom.

AerostructuralDynamics.QuasiSteady — Type

QuasiSteady{Order} <: NoStateModel

2D quasi-steady aerodynamic model with parameters $p_a = \begin{bmatrix} a & b & a_0 & lpha_0 \end{bmatrix}^T$.

AerostructuralDynamics.QuasiSteady — Method

QuasiSteady()

Initialize an object of type QuasiSteady which represents a 2D quasi-steady aerodynamic model.

AerostructuralDynamics.RigidBody — Type

RigidBody <: AbstractModel

Six-degree of freedom rigid body dynamic model with state variables $q = \begin{bmatrix} x & y & z & \phi & \theta & \psi & u & v & w & p & q & r \end{bmatrix}^T$, inputs $r = \begin{bmatrix} F_x & F_y & F_z & M_x & M_y & M_z \end{bmatrix}^T$ and parameters $p = \begin{bmatrix} m & I_{xx} & I_{yy} & I_{zz} & I_{xz} & I_{xy} & I_{yz} \end{bmatrix}^T$.

AerostructuralDynamics.TypicalSection — Type

TypicalSection <: AbstractModel

Typical section structural model with state variables $q = \begin{bmatrix} h & θ & \dot{h} & \dot{\theta} \end{bmatrix}^T$, structural parameters $p_s = \begin{bmatrix} k_h & k_\theta & m & S_\theta & I_\theta \end{bmatrix}^T$, and aerodynamic loads $r = \begin{bmatrix} L & M \end{bmatrix}^T$

AerostructuralDynamics.Wagner — Type

Wagner{TF} <: AbstractModel

Aerodynamic model based on Wagner's function with state variables $\lambda = \begin{bmatrix} \lambda_1 & \lambda_2 \end{bmatrix}^T$, inputs $d = \begin{bmatrix} u & v & \omega \end{bmatrix}^T$ and parameters $p_a = \begin{bmatrix} a & b & a_0 & lpha_0 \end{bmatrix}^T$

AerostructuralDynamics.Wagner — Method

Wagner(; C1=0.165, C2=0.335, eps1 = 0.0455, eps2 = 0.3)

Initialize an object of type Wagner

AerostructuralDynamics.Steady — Method

Steady()

Initialize an object of type QuasiSteady which represents a 2D steady aerodynamic model.

AerostructuralDynamics.get_input_jacobian — Method

get_input_jacobian(models)
get_input_jacobian(models, u, y, p, t)

Calculate the jacobian with respect to the inputs for the specified model or models.

AerostructuralDynamics.get_inputs! — Method

get_inputs!(y, models::NTuple{N,AbstractModel}, u, p, t) where N

In-place version of get_inputs

AerostructuralDynamics.get_inputs — Method

get_inputs(models, u, p, t)

Calculate the inputs to the specified combination of models.

AerostructuralDynamics.get_mass_matrix! — Method

get_mass_matrix!(M, models)
get_mass_matrix!(M, models, u, y, p, t)

In-place version of get_mass_matrix.

AerostructuralDynamics.get_mass_matrix — Method

get_mass_matrix(models)
get_mass_matrix(models, u, y, p, t)

Calculate the mass matrix for a model or combination of models.

AerostructuralDynamics.get_rates! — Method

get_rates!(du, models, u, y, p, t)

In-place version of get_rates

AerostructuralDynamics.get_rates — Method

get_rates(models, u, y, p, t)

Calculate the (mass matrix multiplied) state rates for the specified model or models.

AerostructuralDynamics.get_state_jacobian! — Method

get_state_jacobian!(J, models, u, y, p, t)

In-place version of get_state_jacobian

AerostructuralDynamics.get_state_jacobian — Method

get_state_jacobian(models, u, y, p, t)

Calculate the jacobian with respect to the state variables for the specified models.

AerostructuralDynamics.input_indices — Method

input_indices(models)

Return the indices corresponding to the input variables for each model in models

AerostructuralDynamics.number_of_inputs — Function

number_of_inputs(models)

Return the total number of inputs corresponding to the model or models.

AerostructuralDynamics.number_of_parameters — Function

number_of_parameters(models)

Return the total number of parameters corresponding to the model or models.

AerostructuralDynamics.number_of_states — Function

number_of_states(models)

Return the total number of states corresponding to the model or models.

AerostructuralDynamics.parameter_indices — Method

parameter_indices(models)

Return the indices corresponding to the parameters for each model in models

AerostructuralDynamics.state_indices — Method

state_indices(models)

Return the indices corresponding to the state variables for each model in models