AerostructuralDynamics.AbstractModel — Type

AbstractModel

Supertype for all models.

AerostructuralDynamics.LiftingLineSection — Type

LiftingLineSection <: AbstractModel

Lifting line section model with state variables $v_x, v_y, v_z, \omega_x, \omega_y, \omega_z$, inputs $F_x', F_y', F_z', M_x, M_y, M_z$, and zero parameters. Two-dimensional aerodynamic models may be extended to three-dimensional models by coupling with this model. Note that this model has no rate equations of its own since its state variables are defined as functions of the 3D structural model's state variables.

AerostructuralDynamics.NoStateModel — Type

NoStateModel <: AbstractModel

Supertype for all models which contain no state variables.

AerostructuralDynamics.get_input_jacobian — Method

get_input_jacobian(models)
get_input_jacobian(models, u, y, p, t)

Calculate the jacobian with respect to the inputs for the specified model or models.

AerostructuralDynamics.get_input_mass_matrix — Function

get_input_mass_matrix(models)
get_input_mass_matrix(models, u, p, t)

Calculate the input function mass matrix for the specified combination of models.

AerostructuralDynamics.get_input_mass_matrix! — Function

get_input_mass_matrix!(My, models)
get_input_mass_matrix!(My, models, u, p, t)

In-place version of get_input_mass_matrix.

AerostructuralDynamics.get_input_state_jacobian — Function

get_input_state_jacobian(models::NTuple{N,AbstractModel}, u, p, t) where N

Calculate the jacobian of the input function with respect to the state variables for the specified combination of models.

AerostructuralDynamics.get_input_state_jacobian! — Function

get_input_state_jacobian!(J, models, u, p, t)

In-place version of get_input_state_jacobian

AerostructuralDynamics.get_inputs_from_state_rates — Method

get_inputs_from_state_rates(models, u, p, t)

Calculate the portion of the inputs which are dependent on the state rates. This function is used to test the mass matrices associated with each input mass matrix.

AerostructuralDynamics.get_lhs — Method

get_lhs(model, u, y, p, t)

Return the left hand side of the governing differential equations for the model. This method is used for testing the mass matrix associated with each model.

AerostructuralDynamics.inplaceness — Method

inplaceness(::Type{T})

Return InPlace() if functions associated with model T are in-place or OutOfPlace() if functions associated with model T are out-of-place.

AerostructuralDynamics.inplaceness — Method

inplaceness(::Type{T1}, ::Type{T2}, ..., ::Type{TN})

Return InPlace() if the functions associated with the input function for coupled models T1, T2, ... TN are in-place or OutOfPlace() if the functions associated with the input function for coupled models T1, T2, ... TN are out-of-place.

AerostructuralDynamics.input_jacobian_type — Method

input_jacobian_type(::Type{T})

Return

Empty(), if the jacobian of the mass matrix multiplied state rates with respect to the inputs is empty for model T

Zeros(), if the jacobian of the mass matrix multiplied state rates with respect to the inputs is filled with zeros for model T

Identity(), if the jacobian of the mass matrix multiplied state rates with respect to the inputs is the identity matrix for model T

Constant(), if the jacobian of the mass matrix multiplied state rates with respect to the inputs is constant with respect to time for model T

Linear(), if the jacobian of the mass matrix multiplied state rates with respect to the inputs may vary with respect to time for model T, and is linear with respect to the inputs

Nonlinear(), if the jacobian of the mass matrix multiplied state rates with respect to the inputs may vary with respect to time for model T, and is nonlinear with respect to the inputs

If no method is defined for the specified type, return Nonlinear().

AerostructuralDynamics.mass_matrix_type — Method

mass_matrix_type(::Type{T})

Return

Empty(), if the mass matrix associated with model T is empty

Zeros(), if the mass matrix associated with model T is filled with zeros

Identity(), if the mass matrix associated with model T is the identity matrix

Constant(), if the mass matrix associated with model T is constant with respect to time

Linear(), if the mass matrix associated with model T may vary with respect to time

If no method is defined for the specified type, return Linear().

AerostructuralDynamics.state_jacobian_type — Method

state_jacobian_type(::Type{T})

Return

Empty(), if the jacobian of the mass matrix multiplied state rates with respect to the state variables associated with model T is empty

Zeros(), if the jacobian of the mass matrix multiplied state rates with respect to the state variables associated with model T is filled with zeros

Identity(), if the jacobian of the mass matrix multiplied state rates with respect to the state variables associated with model T is the identity matrix

Constant(), if the jacobian of the mass matrix multiplied state rates with respect to the state variables associated with model T is constant with respect to time

Linear(), if the jacobian of the mass matrix multiplied state rates with respect to the state variables associated with model T may vary with respect to time, but is linear with respect to the states

Nonlinear(), if the jacobian of the mass matrix multiplied state rates with respect to the state variables associated with model T may vary with respect to time, and is nonlinear with respect to the states

If no method is defined for the specified type, return Nonlinear().