Aeroelasticity.CoupledModel — Type

CoupledModel{M,C}

Base type for coupled models in Aeroelasticity.jl.

Aeroelasticity.CoupledModel — Method

CoupledModel(submodels, coupling)

Define a coupled model given a tuple of submodels (of type Submodel) and a coupling model (of type Coupling).

SciMLBase.DAEFunction — Type

DAEFunction(model::Aeroelasticity.CoupledModel, p=nothing)

Construct an DAEFunction for a coupled model which may be used with DifferentialEquations.

SciMLBase.ODEFunction — Type

ODEFunction(model::Aeroelasticity.CoupledModel, p=nothing)

Construct an ODEFunction for a coupled model which may be used with DifferentialEquations.

Aeroelasticity.additional_parameter_indices — Method

additional_parameter_indices(model::CoupledModel)

Return the indices corresponding to additional parameters introduced by the model coupling

Aeroelasticity.get_coupling_inputs! — Method

get_coupling_inputs!(y, models::CoupledModel, dx, x, p, t)

In-place version of get_coupling_inputs

Aeroelasticity.get_coupling_inputs — Method

get_coupling_inputs(model::CoupledModel, dx, x, p, t)

Calculate the coupling inputs for a coupled model.

Aeroelasticity.get_coupling_parameter_jacobian! — Method

get_coupling_parameter_jacobian!(J, model)
get_coupling_parameter_jacobian!(J, model, p)
get_coupling_parameter_jacobian!(J, model, dx, x, p, t)

In-place version of get_coupling_parameter_jacobian

Aeroelasticity.get_coupling_parameter_jacobian — Method

get_coupling_parameter_jacobian(model)
get_coupling_parameter_jacobian(model, p)
get_coupling_parameter_jacobian(model, dx, x, y, p, t)

Calculate the jacobian of the coupling function for model with respect to the parameters

Aeroelasticity.get_coupling_rate_jacobian! — Method

get_coupling_rate_jacobian!(J, model)
get_coupling_rate_jacobian!(J, model, p)
get_coupling_rate_jacobian!(J, model, dx, x, p, t)

In-place version of get_coupling_rate_jacobian.

Aeroelasticity.get_coupling_rate_jacobian — Method

get_coupling_rate_jacobian(model)
get_coupling_rate_jacobian(model, p)
get_coupling_rate_jacobian(model, dx, x, p, t)

Calculate the jacobian of the coupling function with respect to the state rates

Aeroelasticity.get_coupling_state_jacobian! — Method

get_coupling_state_jacobian!(J, model)
get_coupling_state_jacobian!(J, model, p)
get_coupling_state_jacobian!(J, model, dx, x, p, t)

In-place version of get_coupling_state_jacobian

Aeroelasticity.get_coupling_state_jacobian — Method

get_coupling_state_jacobian(model)
get_coupling_state_jacobian(model, p)
get_coupling_state_jacobian(model, dx, x, y, p, t)

Calculate the jacobian of the coupling function for model with respect to the state variables

Aeroelasticity.get_coupling_time_gradient! — Method

get_coupling_time_gradient!(dT, model)
get_coupling_time_gradient!(dT, model, p)
get_coupling_time_gradient!(dT, model, dx, x, p, t)

In-place version of get_coupling_time_gradient

Aeroelasticity.get_coupling_time_gradient — Method

get_coupling_time_gradient(model)
get_coupling_time_gradient(model, p)
get_coupling_time_gradient(model, dx, x, y, p, t)

Calculate the derivative of the coupling function for model with respect to time

Aeroelasticity.get_input_jacobian! — Method

get_input_jacobian!(J, model)
get_input_jacobian!(J, model, p)
get_input_jacobian!(J, model, dx, x, y, p, t)

In-place version of get_input_jacobian

Aeroelasticity.get_input_jacobian — Method

get_input_jacobian(model)
get_input_jacobian(model, p)
get_input_jacobian(model, dx, x, y, p, t)

Calculate the jacobian of the residual function for model with respect to the input variables

Aeroelasticity.get_parameter_jacobian! — Method

get_parameter_jacobian!(J, model)
get_parameter_jacobian!(J, model, p)
get_parameter_jacobian!(J, model, dx, x, y, p, t)

In-place version of get_parameter_jacobian

Aeroelasticity.get_parameter_jacobian — Method

get_parameter_jacobian(model)
get_parameter_jacobian(model, p)
get_parameter_jacobian(model, dx, x, y, p, t)

Calculate the jacobian of the residual function for model with respect to the parameters

Aeroelasticity.get_rate_jacobian! — Method

get_rate_jacobian!(J, model)
get_rate_jacobian!(J, model, p)
get_rate_jacobian!(J, model, dx, x, y, p, t)

In-place version of get_rate_jacobian.

Aeroelasticity.get_rate_jacobian — Method

get_rate_jacobian(model)
get_rate_jacobian(model, p)
get_rate_jacobian(model, dx, x, y, p, t)

Calculate the jacobian of the residual function for model with respect to the state rates

Aeroelasticity.get_residual! — Method

get_residual!(resid, model, dx, x, y, p, t)

In-place version of get_residual

Aeroelasticity.get_residual — Method

get_residual(model, dx, x, y, p, t)

Calculate the residual for the specified model.

Aeroelasticity.get_state_jacobian! — Method

get_state_jacobian!(J, model)
get_state_jacobian!(J, model, p)
get_state_jacobian!(J, model, dx, x, y, p, t)

In-place version of get_state_jacobian

Aeroelasticity.get_state_jacobian — Method

get_state_jacobian(model)
get_state_jacobian(model, p)
get_state_jacobian(model, dx, x, y, p, t)

Calculate the jacobian of the residual function for model with respect to the state variables

Aeroelasticity.get_time_gradient! — Method

get_time_gradient!(dT, model)
get_time_gradient!(dT, model, p)
get_time_gradient!(dT, model, dx, x, y, p, t)

In-place version of get_time_gradient

Aeroelasticity.get_time_gradient — Method

get_time_gradient(model)
get_time_gradient(model, p)
get_time_gradient(model, dx, x, y, p, t)

Calculate the derivative of the residual function for model with respect to time

Aeroelasticity.number_of_additional_parameters — Function

number_of_additional_parameters(model)

Return the total number of additional parameters corresponding to a coupled model.

Aeroelasticity.rate_indices — Method

rate_indices(model::CoupledModel)

Return the rate variable indices for each submodel in model

Aeroelasticity.separate_inputs — Function

separate_inputs(model, y)

Separate the input vector entries in y

Aeroelasticity.separate_rates — Method

separate_rates(model, dx)

Separate the rate vector entries in dx.

Aeroelasticity.set_additional_parameters! — Method

set_additional_parameters!(p, model::CoupledModel; kwargs...)

Set the elements of the parameter vector p corresponding to additional parameters introduced by the model coupling to the values specified in kwargs

Aeroelasticity.set_additional_parameters — Method

set_additional_parameters(model::CoupledModel; kwargs...)

Set the additional parameter vector values for model using the values in kwargs.

Aeroelasticity.set_inputs! — Method

set_inputs!(y, model::CoupledModel, i; kwargs...)

Set the elements of the input vector y corresponding to the ith submodel of model to the values in kwargs

Aeroelasticity.set_inputs! — Method

set_inputs!(y, submodel::Submodel; kwargs...)

In-place version of set_inputs

Aeroelasticity.set_inputs — Method

set_inputs(submodel::Submodel; kwargs...)

Set the input vector values for submodel using the values in kwargs.

Aeroelasticity.set_parameters! — Method

set_parameters!(p, model::CoupledModel, i; kwargs...)

Set the elements of the parameter vector p corresponding to the ith submodel of model to the values in kwargs

Aeroelasticity.set_parameters! — Method

set_parameters!(p, submodel::Submodel; kwargs...)

In-place version of set_parameters

Aeroelasticity.set_parameters — Method

set_parameters(coupling::Coupling; kwargs...)

Set the additional parameter vector values for coupling using the values in kwargs.

Aeroelasticity.set_parameters — Method

set_parameters(submodel::Submodel; kwargs...)

Set the parameter vector values for submodel using the values in kwargs.

Aeroelasticity.set_rates! — Method

set_rates!(dx, submodel::Submodel; kwargs...)

In-place version of set_rates

Aeroelasticity.set_rates — Method

set_rates(submodel::Submodel; kwargs...)

Set the rate vector values for submodel using the values in kwargs.

Aeroelasticity.set_states! — Method

set_states!(x, model::CoupledModel, i; kwargs...)

Set the elements of the state vector x corresponding to the ith submodel of model to the values in kwargs

Aeroelasticity.set_states! — Method

set_states!(x, submodel::Submodel; kwargs...)

In-place version of set_states

Aeroelasticity.set_states — Method

set_states(submodel::Submodel; kwargs...)

Set the state vector values for submodel using the values in kwargs.